+++ dotnet/TF2SA.Http/LogsTF/Service/ILogsTFService.cs
@@ -7,7 +7,7 @@ namespace TF2SA.Http.LogsTF.Service;
 
 public interface ILogsTFService
 {
-	public Task<EitherStrict<HttpError, GameLog>> GetGameLog(uint logId);
+	public Task<EitherStrict<HttpError, GameLog>> GetGameLog(ulong logId);
 	public Task<EitherStrict<HttpError, LogListResult>> GetLogList(LogListQueryParams filter);
 	public Task<EitherStrict<HttpError, List<LogListItem>>> GetAllLogs();
 }
